+++ .npmignore
@@ -1,0 +1,6 @@
+.github
+__tests__
+tsconfig.json
+vitest.config.ts
+.tool-versions
+.prettierignore
