+++ .prettierignore
@@ -1,2 +1,2 @@
 package-lock.json
-src/
+__tests__/fixtures/
