+++ README.md
@@ -3,7 +3,7 @@
 ## Usage
 
 ```bash
-npm i @caspeco/eslint-config
+npm install --save-dev @caspeco/eslint-config eslint@^9 typescript-eslint@^8 @typescript-eslint/parser@^8 @typescript-eslint/eslint-plugin@^8 eslint-config-prettier@^9 eslint-plugin-check-file@^2.8.0 @eslint/js@^9
 ```
 
 The provided configuration contains shared ESLint rules for typescript projects across Caspeco. To use this configuration you'll need `.eslintrc.js` file in your project root with the following content:
