+++ __tests__/fixtures/vanilla/frontend.ts
