+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "@caspeco/eslint-config",
-	"version": "2.0.0-alpha.3",
-	"main": "index.js",
+	"version": "2.0.0-alpha.4",
+	"main": "src/index.js",
 	"type": "module",
 	"homepage": "https://github.com/caspeco/eslint-config#readme",
 	"bugs": {
