+++ src/frontend-vanilla.js
