+++ tsconfig.json
@@ -20,5 +20,5 @@
 		"noUnusedParameters": true,
 		"noFallthroughCasesInSwitch": true
 	},
-	"include": ["src/*.ts", "src/*.tsx", "__tests__/*.ts"]
+	"include": ["src/**/*.js", "__tests__/**/*.ts", "__tests__/**/*.tsx"]
 }
